final_project/web/requirements.txt
flask
flask-cors
kafka-python
